% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{avgReplicates}
\alias{avgReplicates}
\title{Average Replicates in Peptide data frame}
\usage{
avgReplicates(
  df,
  target.samples,
  disease.samples,
  reference.samples,
  chop.length = 2,
  null.ip.col = "Bead",
  split.by = NULL
)
}
\arguments{
\item{df}{phage peptide count matrix}

\item{target.samples}{vector of target samples of interest from one patient (defined by 'patient' input)}

\item{disease.samples}{vector of all disease samples in input data frame}

\item{reference.samples}{vector of all reference samples (healthy and bead controls) in input data frame}

\item{chop.length}{number characters to remove from the end of each sample name to remove the replicate IDs <default: 2>}

\item{null.ip.col}{option that indicates a regex to group all AG Bead columns <default: Bead >}

\item{split.by}{insert regex to split sample strings by and keep the first element, which helps simplify complex sample names <default: NULL>}
}
\value{
data frame with non-target columns averaged
}
\description{
takes input of phage results used for Heatmaps and averages the replicates on non=target columns
}
\examples{
avgReplicates(df,target.samples,disease.samples,reference.samples)
}
