% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZscores.R
\name{calcZscores}
\alias{calcZscores}
\title{Creates Z-score Columns}
\usage{
calcZscores(df, target.samples, reference.samples, method = "mean", MIN = 1)
}
\arguments{
\item{df}{phage peptide count matrix}

\item{target.samples}{vector of target samples to analyze}

\item{reference.samples}{vector of reference control samples used for calculating Std. Dev. and average expression}

\item{method}{method to use for calculating Z-scores ("median","mean") <default: "mean">}

\item{MIN}{minimum count value for the inputted data frame <default: 1>}
}
\value{
returns phage results with FC columns for each target sample.
}
\description{
takes input of phage results and calculates Z-scores for every target column based on the input
}
\examples{
calcZscores(df, target_columns)
}
