% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullParse.R
\name{filterPeptides}
\alias{filterPeptides}
\title{Simple Filters for phage data}
\usage{
filterPeptides(
  peptide_df,
  list_of_samples,
  RPK_THRESH = 20,
  FC_THRESH = 100,
  peptide.id.col = "peptide_id"
)
}
\arguments{
\item{peptide_df}{phage peptide count matrix}

\item{list_of_samples}{list of all samples. It is assumed the FC columns are sample names with '_FC' at the end}

\item{RPK_THRESH}{RPK threshold <default: 20>}

\item{FC_THRESH}{Fold Change threshold <default: 100>}

\item{peptide.id.col}{Name of column with unique Peptide ID <default: "peptide_id">}
}
\value{
returns a subset of phage results with peptides that pass all filters.
}
\description{
takes input of phage results and applies RPK and Fold Change filters
then returns a subset of phage results. This is more simple version of 'fullParse'
}
\examples{
filterPeptides(peptide_df,list_of_samples, RPK_THRESH = 20,FC_THRESH = 100,peptide.id.col = "peptide_id")
}
