% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullParse.R
\name{findKmerOverlap}
\alias{findKmerOverlap}
\title{Identify Kmer overlaps}
\usage{
findKmerOverlap(peptide_df, KMER_SIZE = 7)
}
\arguments{
\item{peptide_df}{phage peptide count matrix}

\item{KMER_SIZE}{size of Kmer <default: 7>}
}
\value{
returns a subset of phage results with peptides that contain Kmer overlaps.
}
\description{
takes input of phage results and finds all Kmer overlaps among all peptide sequences
then returns a subset of phage results with peptides that contain overlaps
}
\examples{
findKmerOverlap(peptide_df,KMER_SIZE = 7)
}
