% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullParse.R
\name{fullParse}
\alias{fullParse}
\title{Filters phage data}
\usage{
fullParse(
  df,
  list_of_samples,
  MIN_RPK = 0,
  FC_THRESH1 = 10,
  FC_THRESH2 = 100,
  SUM_RPK_THRESH = 50,
  ZSCORE_THRESH1 = 0,
  ZSCORE_THRESH2 = 0
)
}
\arguments{
\item{df}{phage peptide count matrix}

\item{list_of_samples}{list of target samples to analyze}

\item{MIN_RPK}{Threshold for minimum RPK value <default: 0>}

\item{FC_THRESH1}{Fold Change threshold 1 <default: 10>}

\item{FC_THRESH2}{Fold Change threshold 2 <default: 100>}

\item{SUM_RPK_THRESH}{Gene level total RPK threshold <default: 50>}

\item{ZSCORE_THRESH1}{If calculated, applies minimum Z-score threshold for each peptide <default: 0>}

\item{ZSCORE_THRESH2}{If calculated, applies additional Z-score threshold similar to the FC_THRESH2 arguement <default: 0>}
}
\value{
returns a subset of phage results with peptides that pass all filters.
}
\description{
takes input of phage results and applies RPK and Fold Change filters
then returns a subset of phage results
}
\examples{
fullParse(df,list_of_samples, MIN_RPK = 2,FC_THRESH1 = 10, FC_THRESH2 = 100, SUM_RPK_THRESH = 50)
}
