% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFC.R
\name{makeFCdf}
\alias{makeFCdf}
\title{Creates Fold Change Columns}
\usage{
makeFCdf(df, target_columns, mean_column)
}
\arguments{
\item{df}{phage peptide count matrix}

\item{target_columns}{list of target samples to analyze}

\item{mean_column}{column containing the mean RPK of samples that represent the controls for normalization}
}
\value{
returns phage results with FC columns for each target sample.
}
\description{
takes input of phage results and calculates Fold Change based on input columns
}
\examples{
makeFCdf(df, target_columns, mean_column)
}
