% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Heatmap of Peptide Hits}
\usage{
makeHeatmap(
  df,
  patient,
  gene.col = "gene",
  peptide.id.col = "peptide_id",
  target.samples,
  disease.samples,
  reference.samples,
  normalization.method = "log10",
  disease.annot = "CoV",
  reference.annot = "Reference"
)
}
\arguments{
\item{df}{phage peptide count matrix}

\item{patient}{patient of interest}

\item{gene.col}{column name for gene name in input data frame <default: "gene">}

\item{peptide.id.col}{column name for peptide IDs in input data frame <default: "peptide_id">}

\item{target.samples}{vector of target samples of interest from one patient (defined by 'patient' input)}

\item{disease.samples}{vector of all disease samples in input data frame}

\item{reference.samples}{vector of all reference samples (healthy and bead controls) in input data frame}

\item{normalization.method}{method used to normalize data <'log10', 'minmax','meannorm'>}

\item{disease.annot}{column annotation for disease samples}

\item{reference.annot}{column annotation for reference samples}
}
\value{
variable containing resulting heatmap.
}
\description{
takes input of phage results and generates a heatmap of all peptides
then returns variable containing the plot
}
\examples{
makeHeatmap(df,patient,target.samples,disease.samples,reference.samples)
}
