% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZscores.R
\name{calcZscores}
\alias{calcZscores}
\title{Creates Z-score Columns}
\usage{
calcZscores(
  df,
  target.samples,
  method = "median",
  MIN = 1,
  peptide.id.col = "peptide_id",
  onlyZscores = FALSE
)
}
\arguments{
\item{df}{phage peptide count matrix}

\item{method}{method to use for calculating Z-scores ("median","mean") <default: "median">}

\item{MIN}{minimum count value for the inputted data frame <default: 1>}

\item{onlyZscores}{option to format the output and return full data frame or only the Z-score columns <default: FALSE>}

\item{target.columns}{list of target samples to analyze}
}
\value{
returns phage results with FC columns for each target sample.
}
\description{
takes input of phage results and calculates Z-scores for every target column based on the input
}
\examples{
calcZscores(df, target_columns)
}
