% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{collapseReplicateColumnNames}
\alias{collapseReplicateColumnNames}
\title{Format Column names to remove replicate IDs}
\usage{
collapseReplicateColumnNames(col.name.vec, chop.length = 2)
}
\arguments{
\item{col.name.vec}{input column names, sample names only}

\item{chop.length}{number of characters to chop from the end of each sample name <default: 2>}
}
\value{
vector of unique collapsed sample names
}
\description{
take an input vector of sample names and collapse all replicates
}
\examples{
collapseReplicateColumnNames(col.name.vec,chop.length = 2)
}
