% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findKmerOverlap.R
\name{findKmerOverlap}
\alias{findKmerOverlap}
\title{Identify Kmer overlaps}
\usage{
findKmerOverlap(
  peptide_df,
  KMER_SIZE = 7,
  TEMP_FASTA = "./temp_pep.fasta",
  seq.col = "sequence",
  peptide.col = "peptide_id"
)
}
\arguments{
\item{peptide_df}{phage peptide count matrix}

\item{KMER_SIZE}{size of Kmer <default: 7>}

\item{TEMP_FASTA}{file handle for a temporary FASTA that needs to be generated in order to perform the Kmer analysis <default: "./temp_pep.fasta">}

\item{seq.col}{column name for peptide AA sequences in input data frame <default: "sequence">}

\item{peptide.col}{column name for peptide IDs in input data frame <default: "peptide_id">}
}
\value{
returns a subset of phage results with peptides that contain Kmer overlaps.
}
\description{
takes input of phage results and finds all Kmer overlaps among all peptide sequences
then returns a subset of phage results with peptides that contain overlaps
}
\examples{
findKmerOverlap(peptide_df,KMER_SIZE = 7,seq.col = "sequence",peptide.col = "peptide_id")
}
