% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Heatmap of Peptide Hits}
\usage{
makeHeatmap(
  df,
  patient,
  gene.col = "gene",
  peptide.id.col = "peptide_id",
  target.samples,
  disease.samples,
  reference.samples,
  normalization.method = "log10",
  disease.annot = "COVID-19+",
  reference.annot = "REFERENCE",
  min.rpk = 0,
  break.list = NULL,
  gene.order = c(),
  avg.non.target.columns = FALSE,
  disease.first = TRUE,
  cell.height = 5,
  cell.width = 5,
  num.pos.controls = 1,
  null.ip.col = "Bead",
  transpose.heatmap = FALSE,
  custom.gene.labels = TRUE
)
}
\arguments{
\item{df}{phage peptide count matrix}

\item{patient}{patient of interest}

\item{gene.col}{column name for gene name in input data frame <default: "gene">}

\item{peptide.id.col}{column name for peptide IDs in input data frame <default: "peptide_id">}

\item{target.samples}{vector of target samples of interest from one patient (defined by 'patient' input)}

\item{disease.samples}{vector of all disease samples in input data frame}

\item{reference.samples}{vector of all reference samples (healthy and bead controls) in input data frame}

\item{normalization.method}{method used to normalize data <'log10', 'minmax','meannorm'>}

\item{disease.annot}{column annotation for disease samples}

\item{reference.annot}{column annotation for reference samples}

\item{min.rpk}{For heatmaps set 0 values to an inputted minimum RPK value <default: 0>}

\item{break.list}{This determines the margins for the resultsing heatmap. the only options are (1,2,3,4) <default: NULL>}

\item{gene.order}{Vector of gene names to force the order in which peptides are displayed in the heatmap <default: c()>}

\item{avg.non.target.columns}{Average the replicates for all samples that are not in target.samples <default: FALSE>}

\item{disease.first}{option for where disease group apears first on the heatmap <default: TRUE>}

\item{cell.height}{option for cell height in heatmap <default: 5>}

\item{cell.width}{option for cell width in heatmap <default: 5>}

\item{num.pos.controls}{option to indicate number of AG Bead controls in the data frame, if you are averaging replicates, keep this number at 1 <default: 1>}

\item{null.ip.col}{option that indicates a regex to group all AG Bead columns <default: Bead >}

\item{transpose.heatmap}{option to transpose the heatmaps columns and rows <default: FALSE>}

\item{custom.gene.labels}{option to change row labels to gene names only labeled once per gene block <default: TRUE>}
}
\value{
variable containing resulting heatmap.
}
\description{
takes input of phage results and generates a heatmap of all peptides
then returns variable containing the plot
}
\examples{
makeHeatmap(df,patient,target.samples,disease.samples,reference.samples)
}
